canonicalize can no longer be piped to other commands
The purpose of c14n is to produce a string, not quads,
so it is cleaner to restrict this subcommand to be the final one.

People may still pipe the result into a new instance of sop if they want,
but there is no reason to make the code of sop more complex,
just to handle this niche use case.