use my own variant of the clap Verbosity option
the original one is marked as "global",
which means it appears under the help section of all subcommands as well.
However, this was confusing,
because it was only actually supported in the *first* subcommand of the pipeline.

I decided that it was better to require it before the first subcommand,
to avoid any confusion.