do not propagate errors when parsing multiple files
previously, when parsing multiple files,
any error was propagated down the pipeline,
causing the consumer to stop early and close the MPSC channel,
causing in turn an error in the other parser threads.

now, when --multiple is used,
any error encountered by a parser is logged as a warning,
but not propagated down the pipeline.