more consistent handling of formats
the format for the 'serializer' command is now optional (named --format/-f),
and is guessed from the output filename (--output/-o) when present.

Also, the --format option overrides "format guessing"
(from file extension or HTTP content-type)
when applied to a single argument (single source 'parse', or 'serialize')
but is only used when guessing fails
when applied to multiple arguments (multi-source 'parse', or 'dispatch').

This seems to be a sensible default.