add support for alternative c14n Sophia-C14N
Sophia-C14N is a non-standard extension of RDFC-1.0,
which supports RDF 1.2 and most of Sophia's Generalized RDF.
It is therefore useful to compare RDF files beyond strict RDF 1.1.

Note however that Sophia-C14N may change in future versions of Sophia,
so it should not be used for, e.g., digital signatures.